Log config autofix for missing max position size
